+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added
   - Support for method-binding patterns e.g. `this.foo = this.foo.bind(this);`
+  - Support for multiple states in one `setState()` call
 
 ## 0.1.8